static関数のファイル分割を実施
